+++ src/pages/Community/Notice/NoticeSection.styled.ts
@@ -74,7 +74,7 @@ export const BoardHeaderContainer = styled.section`
   flex-direction: column;
 `;
 
-const NoticeBoardRowStyles = css`
+export const NoticeBoardRowStyles = css`
   width: 100%;
   height: fit-content;
   padding-right: 4.4rem;
@@ -116,7 +116,7 @@ export const NoticeBoardCredit = styled.div`
   align-items: center;
 `;
 
-const creditWidthStyles = css`
+export const creditWidthStyles = css`
   flex: 1;
   text-align: center;
 `;
@@ -131,85 +131,6 @@ export const AuthorTitle = styled.span`
 
 export const AttatchmentTitle = styled.span`
   ${creditWidthStyles}
-`;
-
-// 게시글 Section
-export const NoticeBoardPostRow = styled.div`
-  ${NoticeBoardRowStyles}
-  margin: 3.2rem 0;
-`;
-
-export const PostTitleContainer = styled.div`
-  width: 60%;
-  height: fit-content;
-`;
-
-export const PostLink = styled.a`
-  width: fit-content;
-  height: fit-content;
-
-  display: flex;
-  align-items: center;
-  gap: 0.8rem;
-`;
-
-// 중요 공지 Box 우측 Divider
-export const TinyDivider = styled.div`
-  width: 1px;
-  height: 1.4rem;
-
-  background-color: ${({ theme }) => theme.colors.HID_Grayscale[200]};
-`;
-
-// 게시글 제목
-export const PostTitle = styled.span`
-  ${({ theme }) => theme.fontStyles.Body3_KO}
-  color: ${({ theme }) => theme.colors.HID_Grayscale[800]};
-
-  overflow: hidden;
-  text-overflow: ellipsis;
-`;
-
-export const UploadDate = styled.span`
-  ${creditWidthStyles}
-
-  ${({ theme }) => theme.fontStyles.Body3_KO}
-  color: ${({ theme }) => theme.colors.HID_Grayscale[500]};
-`;
-
-export const Author = styled.span`
-  ${creditWidthStyles}
-`;
-
-export const Attatchment = styled.span`
-  ${creditWidthStyles}
-`;
-
-// Author Box 영역
-export const AuthorBoxArea = styled.div`
-  flex: 1;
-
-  display: flex;
-  justify-content: center;
-  align-items: center;
-`;
-
-// 첨부파일 영역
-export const ClipIconContainer = styled.div`
-  flex: 1;
-
-  display: flex;
-  justify-content: center;
-  align-items: center;
-`;
-
-export const ClipIconButton = styled.button`
-  width: 2.4rem;
-  height: 2.4rem;
-
-  display: flex;
-  justify-content: center;
-  align-items: center;
 `;
 
 // 게시글 하단 Divider (Unit)
