+++ src/styles/global/_normalize.scss
@@ -58,6 +58,10 @@ dd {
   margin: 0;
 }
 
+input {
+  outline: none;
+}
+
 // 5. 리스트 스타일 정리
 ul,
 ol {
